Added new words
RAND: push a random number onto the stack
BUF: pushes address of string buffer onto the stack
WORD: read the next word from the user
PUTS: display a null-terminated string at the address given to by TOS
PUTLN: same as PUTS, but with a newline
TO_ASCII: converts the top of the stack (assumed to be a keycode) to the corresponding ASCII code.

Still need to document everything!